Make Package

Create a new R package with the specified parameters

mk_package(
  path,
  author,
  email,
  git_username = NULL,
  git_email = NULL,
  readme_md = TRUE,
  git = TRUE,
  check_pkg_name = TRUE,
  ci = "travis",
  license = "MIT",
  pkgdown = TRUE
)

Arguments

path

character. The path to create the package at.

author

character. The name of the package author.

email

character. The email to use for Git and GitHub.

git_username

character. The username to use for Git and GitHub.

readme_md

logical. Whether to create a README file in markdown format.

git

logical. Whether to initialize a Git repository for the package.

check_pkg_name

logical. Whether to check that the package name is available on CRAN.

ci

character. The continuous integration service to use (either "travis" or "appveyor").

license

character. The license to use for the package (either "MIT" or "GPL-3").

pkgdown

logical. Whether to create a pkgdown site for the package.